Move 'predict()' function into predictor class
