+++ tests/test_prediction.py
@@ -5,12 +5,12 @@ from pathlib import Path
 import pytest
 import yaml
 from clr_loader import get_mono
-from spectrseqtools.cli import predict
 from spectrseqtools.dataclasses import Sequence
 from spectrseqtools.enums import SolverType
 from spectrseqtools.nucleotide_alphabet import NucleotideAlphabet
 from spectrseqtools.parsers import PredictionOptions, PreprocessingOptions
 from spectrseqtools.plotting import plot_prediction
+from spectrseqtools.prediction.prediction import Predictor
 from spectrseqtools.preprocessing.preprocessing import Preprocessor
 
 rt = get_mono()
@@ -54,7 +54,7 @@ def test_testcase(testcase):
         with open(base_path / "fragments.preprocessed.meta.yaml", "w") as f:
             yaml.safe_dump(meta, f)
 
-    prediction = predict(
+    prediction = Predictor(
         PredictionOptions(
             fragments=base_path / "fragments.tsv",
             meta=base_path / "fragments.preprocessed.meta.yaml",
@@ -66,7 +66,7 @@ def test_testcase(testcase):
             solver=SolverType.CBC,
             # solver=SolverType.GUROBI,
         )
-    )
+    ).predict()
 
     # Read true sequence from meta file
     true_seq = Sequence.from_str(meta["true_sequence"])
